refactor: Streamline EIP-712 signature validation by moving implementation to a dedicated function and removing redundant code
