+++ crates/solver-order/src/implementations/standards/_7683.rs
@@ -8,7 +8,6 @@ use crate::{OrderError, OrderInterface};
 use alloy_primitives::{Address as AlloyAddress, Bytes, FixedBytes, U256};
 use alloy_sol_types::{SolCall, SolType};
 use async_trait::async_trait;
-use solver_delivery::DeliveryService;
 use solver_types::{
 	current_timestamp,
 	oracle::OracleRoutes,
@@ -22,7 +21,6 @@ use solver_types::{
 	Address, ConfigSchema, Eip7683OrderData, ExecutionParams, FillProof, Intent, NetworksConfig,
 	Order, OrderStatus, Schema, Transaction,
 };
-use std::sync::Arc;
 
 /// EIP-7683 order implementation.
 ///
