+++ crates/solver-service/src/eip712/compact/mod.rs
@@ -39,7 +39,7 @@ impl SignatureValidator for CompactSignatureValidator {
 		signature: &Bytes,
 		expected_signer: AlloyAddress,
 	) -> Result<bool, APIError> {
-		validate_eip712_signature(domain_separator, struct_hash, signature, expected_signer)
+		crate::eip712::validate_eip712_signature(domain_separator, struct_hash, signature, expected_signer)
 	}
 
 	/// Extract sponsor signature from ABI-encoded signature bytes
@@ -252,85 +252,3 @@ pub fn compute_batch_compact_struct_hash(
 	Ok(keccak256(struct_data))
 }
 
-/// Validate EIP-712 signature against expected signer
-pub fn validate_eip712_signature(
-	domain_separator: FixedBytes<32>,
-	struct_hash: FixedBytes<32>,
-	signature: &Bytes,
-	expected_signer: AlloyAddress,
-) -> Result<bool, APIError> {
-	// Compute EIP-712 message hash
-	let message_hash = keccak256(
-		[
-			&[0x19, 0x01][..],
-			domain_separator.as_slice(),
-			struct_hash.as_slice(),
-		]
-		.concat(),
-	);
-
-	// Recover signer from signature
-	let recovered_signer = recover_signer(message_hash, signature)?;
-
-	Ok(recovered_signer == expected_signer)
-}
-
-/// Recover signer address from signature and message hash
-fn recover_signer(
-	message_hash: FixedBytes<32>,
-	signature: &Bytes,
-) -> Result<AlloyAddress, APIError> {
-	if signature.len() != 65 {
-		return Err(APIError::BadRequest {
-			error_type: ApiErrorType::OrderValidationFailed,
-			message: "Invalid signature length".to_string(),
-			details: None,
-		});
-	}
-
-	let recovery_id = signature[64];
-	let recovery_id = if recovery_id >= 27 {
-		recovery_id - 27
-	} else {
-		recovery_id
-	};
-
-	let signature_bytes = &signature[0..64];
-
-	use secp256k1::{ecdsa::RecoverableSignature, All, Message, Secp256k1};
-
-	let secp = Secp256k1::<All>::new();
-
-	// Create recoverable signature
-	let recovery_id = secp256k1::ecdsa::RecoveryId::try_from(recovery_id as i32).unwrap();
-	let recoverable_sig = RecoverableSignature::from_compact(signature_bytes, recovery_id)
-		.map_err(|_| APIError::BadRequest {
-			error_type: ApiErrorType::OrderValidationFailed,
-			message: "Invalid signature format".to_string(),
-			details: None,
-		})?;
-
-	// Create message from hash
-	let message =
-		Message::from_digest_slice(message_hash.as_slice()).map_err(|_| APIError::BadRequest {
-			error_type: ApiErrorType::OrderValidationFailed,
-			message: "Invalid message hash".to_string(),
-			details: None,
-		})?;
-
-	// Recover public key
-	let public_key = secp
-		.recover_ecdsa(&message, &recoverable_sig)
-		.map_err(|_| APIError::BadRequest {
-			error_type: ApiErrorType::OrderValidationFailed,
-			message: "Failed to recover public key".to_string(),
-			details: None,
-		})?;
-
-	// Get address from public key
-	let public_key_bytes = public_key.serialize_uncompressed();
-	let public_key_hash = keccak256(&public_key_bytes[1..]);
-	let address_bytes = &public_key_hash.as_slice()[12..];
-
-	Ok(AlloyAddress::from_slice(address_bytes))
-}
